renamed products folder
