+++ src/app/app-routing.module.ts
@@ -5,7 +5,7 @@ import { DashboardComponent } from './Pages/dashboard/dashboard.component';
 import { LoginComponent } from './Pages/login/login.component';
 import { OrderComponent } from './Pages/order/order.component';
 import { PlumberComponent } from './Pages/plumber/plumber.component';
-import { ProductDisplayComponent } from './Pages/product/products-display.component';
+import { ProductDisplayComponent } from './Pages/product-display/products-display.component';
 import { TicketComponent } from './Pages/ticket/ticket.component';
 
 const routes: Routes = [
